requirements.txt
protobuf==3.4.0
grpcio==1.14.1
grpcio-tools==1.14.1
sphinx==1.7.6
sphinx-click==1.3.0
flask==0.12.2
elasticsearch>=6.0.0,<7.0.0